fractionAsFloat() should obviously return a float...
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@6338 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/FixedPoint.hh
@@ -44,7 +44,7 @@ public:
 	int toInt() const { return value >> FRACTION_BITS; }
 	float toFloat() const { return static_cast<float>(value) / ONE; }
 	double toDouble() const { return static_cast<double>(value) / ONE; }
-	double fractionAsFloat() const {
+	float fractionAsFloat() const {
 		return static_cast<float>(value & FRACTION_MASK) / ONE;
 	}
 	double fractionAsDouble() const {
